+++ Samples~/MainSample/Scripts/Inputs/DualsensePauseHaptics.cs
@@ -1,0 +1,18 @@
+using DualSenseSample.Inputs;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class DualsensePauseHaptics : AbstractDualSenseBehaviour
+{
+    public void PauseHaptics()
+    {
+        if (DualSense != null)
+        {
+            if (!DualSense.IsHapticPaused)
+                DualSense.PauseHaptics();
+            else
+                DualSense.ResumeHaptics();
+        }
+    }
+}